task route protected by jwt

+++ src/Models/taskModel.js
@@ -1,9 +1,9 @@
 const taskModel = {
-    select:`SELECT * FROM tasks`,
-    selectById: `SELECT * FROM tasks WHERE id = ?`,
-    insert: `INSERT INTO tasks (title, description, due_date) VALUES (?, ?, ?)`,
-    update: `UPDATE tasks set title=?, description= ? WHERE id = ?`,
-    deleteById: `DELETE FROM tasks WHERE id = ?`
+    select:`SELECT * FROM tasks WHERE user_id = ?`,
+    selectById: `SELECT * FROM tasks WHERE id = ? AND user_id = ?`,
+    insert: `INSERT INTO tasks (title, description, due_date, user_id) VALUES (?, ?, ?, ?)`,
+    update: `UPDATE tasks set title=?, description= ?, due_date = ? WHERE id = ? AND user_id = ?`,
+    deleteById: `DELETE FROM tasks WHERE id = ? AND user_id = ?`
 }
 
 export { taskModel };
